wrapped the metamask components into a more general metamask component to not clutter app.js
