+++ src/App.js
@@ -1,5 +1,4 @@
 /* eslint-disable camelcase */
-/* eslint-disable class-methods-use-this */
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
@@ -15,26 +14,12 @@ import PortfolioPage from './components/pages/PortfolioPage';
 import TransactionHistoryPage from './components/pages/TransactionHistoryPage';
 import AppHeader from './components/AppHeader';
 import NavigationBar from './components/NavigationBar';
-import MetamaskBooting from './components/MetamaskBooting';
-import MetamaskLogin from './components/MetamaskLogin';
-import BrowserNotSupported from './components/BrowserNotSupported';
+import MetamaskChecker from './components/MetamaskChecker';
 
 import * as actions from './actions';
 import { MYBIT_TICKER_COINMARKETCAP, ETHEREUM_TICKER_COINMARKETCAP } from './constants';
 
-import isMetaMask from './util/isMetamask';
-import checkAccount from './util/isUserLogged';
-
-const { detect } = require('detect-browser');
-
 class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isMetamaskUserLogged: null,
-    };
-  }
-
   UNSAFE_componentWillMount() {
     this.props.fetchPriceFromCoinmarketcap(MYBIT_TICKER_COINMARKETCAP);
     this.props.fetchPriceFromCoinmarketcap(ETHEREUM_TICKER_COINMARKETCAP);
@@ -46,39 +31,6 @@ class App extends Component {
     }, timeout);
   }
 
-  componentDidMount() {
-    checkAccount().then((haveAccounts) => {
-      if (haveAccounts.length === 0) {
-        this.setState({ isMetamaskUserLogged: false });
-      }
-    });
-  }
-
-  isBrowserSupported() {
-    const browser = detect();
-    const supportedBrowsers = [
-      'chrome',
-      'opera',
-      'brave',
-      'firefox',
-    ];
-    if (supportedBrowsers.includes(browser.name)) {
-      return true;
-    }
-    return false;
-  }
-
-  // if Metamask is not established, modal is displayed with directions
-  renderMetamaskWarrning() {
-    if (!this.isBrowserSupported()) {
-      return <BrowserNotSupported />;
-    }
-    if (!isMetaMask()) {
-      return <MetamaskBooting />;
-    }
-    return null;
-  }
-
   render() {
     const { state, setTransactionHistoryFilters, fetchTransactionHistory } = this.props;
 
@@ -88,8 +40,7 @@ class App extends Component {
           state={this.props.state}
         />
         <NavigationBar currentPath={this.props.location.pathname} />
-        {this.renderMetamaskWarrning()}
-        {(this.state.isMetamaskUserLogged === false) ? <MetamaskLogin /> : null }
+        <MetamaskChecker />
         <div className="page-wrapper">
           <Switch>
             <Route exact path="/" component={() => <Redirect to="/explore" />} />
